Added progress for employee detail view last commit, refined it here. Scaffolding changes: Updated create method for employees. Try catch added to handle errors, ID in binding removed as SQL Server determines ID automatically. Edit page updated to implement overposting protection (security) according to Microsoft Tutorials. Delete GET and POST updated to include error messages and error handling and maintain redirection to list of employees. Delete view now has error  messages.

+++ TrainingTracker/Controllers/EmployeesController.cs
@@ -35,8 +35,7 @@ namespace TrainingTracker.Controllers
                 return NotFound();
             }
 
-            //var employee = await _context.Employees
-            //    .FirstOrDefaultAsync(m => m.EmployeeId == id);
+            //Get progress for employees where IDs match.
             var employee = await _context.Employees
              .Include(s => s.Progresses)
                  .ThenInclude(e => e.Training)
@@ -65,16 +64,26 @@ namespace TrainingTracker.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("EmployeeId,FirstName,LastName")] Employee employee)
         {
-            if (ModelState.IsValid)
+            try{
+            
+                if (ModelState.IsValid)
+                {
+                    _context.Add(employee);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException /* ex */)
             {
-                _context.Add(employee);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                //Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
             }
             return View(employee);
         }
 
-        // GET: Employees/Edit/5
+        // GET: Employees/Edit/ID
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -90,43 +99,39 @@ namespace TrainingTracker.Controllers
             return View(employee);
         }
 
-        // POST: Employees/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("EmployeeId,FirstName,LastName")] Employee employee)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != employee.EmployeeId)
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var employeeToUpdate = await _context.Employees.FirstOrDefaultAsync(s => s.EmployeeId == id);
+            if (await TryUpdateModelAsync<Employee>(
+                employeeToUpdate,
+                "",
+                s => s.FirstName, s => s.LastName))
             {
                 try
                 {
-                    _context.Update(employee);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!EmployeeExists(employee.EmployeeId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists, " +
+                        "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(employee);
+            return View(employeeToUpdate);
         }
 
-        // GET: Employees/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+ 
+
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -134,29 +139,48 @@ namespace TrainingTracker.Controllers
             }
 
             var employee = await _context.Employees
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.EmployeeId == id);
             if (employee == null)
             {
                 return NotFound();
             }
 
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
+
             return View(employee);
         }
 
-        // POST: Employees/Delete/5
+
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var employee = await _context.Employees.FindAsync(id);
-            _context.Employees.Remove(employee);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (employee == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Employees.Remove(employee);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
-        private bool EmployeeExists(int id)
-        {
-            return _context.Employees.Any(e => e.EmployeeId == id);
-        }
+
+
     }
 }
